RFCT: extract mutation into api folder and update hooks
