+++ frontend/src/features/evaluation-results/hooks/useEvaluate.js
@@ -1,26 +1,9 @@
 import { useEffect } from "react";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { useShallow } from "zustand/react/shallow";
 
 import useTrainStore from "@/store/train";
-import { startEvaluation } from "@/services/api-ai-model";
-
-const useEvaluationResults = ({ updateError }) => {
-  const queryClient = useQueryClient();
-  return useMutation({
-    mutationKey: ["evaluateModel"],
-    mutationFn: async ({ item, ai_model_name }) =>
-      await startEvaluation(item, { ai_model_name }),
-    onSuccess: (data) => {
-      queryClient.setQueryData(["evaluatedModel"], data);
-    },
-    onError: (error) => {
-      console.log(error.message);
-      updateError(error.message);
-      queryClient.removeQueries(["evaluatedModel"]); // Clear cache on error
-    },
-  });
-};
+import { useEvaluationResults } from "../api/evaluation-results";
 
 const useEvaluate = ({ updateError, item }) => {
   const { status, updateStatus } = useTrainStore(
